rename test-env
