+++ test/suites/example/example.conf.ts
@@ -1,4 +1,4 @@
-import TestEnv from '../../setup/TestEnv.js';
+import TestEnv from '../../setup/test-env.js';
 import wdioConfig from '../../setup/wdio.conf.js';
 
 export const testEnv = TestEnv.createWithDefaults( {
